Fixed for ppc darwin.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31216 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/platform/platform_darwin_runtime_os.c
@@ -140,23 +140,25 @@ char *platform_get_darwin_runtime_cpu(void)
         }
 
 #ifdef __ppc__
-        tempfile = archdep_tmpnam();
-        tempsystem = util_concat("uname -m >", tempfile, NULL);
-        system(tempsystem);
-        infile = fopen(tempfile, "rb");
-        if (infile) {
-            amount = fread(buffer, 1, 4095, infile);
-            if (amount) {
-                buffer[strlen(buffer)] = 0;
-                if (strcmp(buffer, "Power Macintosh")) {
-                    sprintf(cpu, "%s [Rosetta (%s)]", cpu, buffer);
+        if (strncmp(cpu, "ppc", 3)) {
+            tempfile = archdep_tmpnam();
+            tempsystem = util_concat("uname -m >", tempfile, NULL);
+            system(tempsystem);
+            infile = fopen(tempfile, "rb");
+            if (infile) {
+                amount = fread(buffer, 1, 4095, infile);
+                if (amount) {
+                    buffer[strlen(buffer)] = 0;
+                    if (strcmp(buffer, "Power Macintosh")) {
+                        sprintf(cpu, "%s [Rosetta (%s)]", cpu, buffer);
+                    }
                 }
+                fclose(infile);
+                unlink(tempfile);
             }
-            fclose(infile);
-            unlink(tempfile);
+            lib_free(tempsystem);
+            lib_free(tempfile);
         }
-        lib_free(tempsystem);
-        lib_free(tempfile);
 #endif
         got_cpu = 1;
     }
